Add ability to import content into S3
Previously, the `uri` of a POSTed version had to already be in an archive we control. Now we can pull in content from somewhere else and save it to S3. The storage layer (S3) is pluggable, so we can write a store for Google Cloud, a filesystem,  or anywhere else.