Fix totally borked invitation code generation
Ridiculously, the first invitation created would wind up having no invitation code. This is because I'm an idiot and am not used to Ruby's while/if-at-the-end-of-the-line style and didn't realize the code would first be initialized to Nil instead of a generated hash value. Once we had one bad invitation in the DB, others would behave because the routine would see that their Nil code collided with the original Nil code.

Also fixes the poorly-coded fixture that masked this issue in tests.