Restructure diffing as pluggable module
There is a registry of diffing services that can be set with:

    Differ.register(type, service)

And retrieved with:

    Differ.for_type(type) # returns nil for unknown type
    Differ.for_type!(type) # raises for unknown type

A diffing service need only respond to the `diff(change_model, options_hash)` method.

Right now there is only one diffing service: SimpleDiff. It's configurable with a URL and simply requests the URL with the query args:
- `a`: the uri for the content of the start version
- `b`: the uri for the content of the end version