Return 404 error when annotating first version
The first version of a page can't be annotated (because annotations are really on the change between a version and its previous version -- and there is no version before the first version). We raised an exception from down in the depths of the app for this before, but this provides clients with a 4XX error (indicating the issue is on their side) and a clearer message about what was wrong (no previous version to annotate).

Fixes #81