Ensure unique pages if querying by version attrs
Joining pages to versions means our output will have repeated pages. The API, however, should be returning a unique list of pages (with versions rolled up as attributes of the pages). This is pretty similar to #85.

The `distinct` call would be more technically correct if only used where we do a join, but always adding at the end like this shouldn't have major negative effects and protects against future mistakes if we add joins elsewhere.