Fix missing pages when include_versions is true
We previously did not include all the matching pages when paging through the results for a large query on a list of pages (two kinds of "pages" there, con't figure out a better way to phrase). This is because of some craziness with how ordering is handled when using limit/offset queries in ActiveRecord.