Fix previous ordering/paging bug better
Instead of sorting in Ruby code, emulate the two-step query ActiveRecord automatically does (but work around the includes/order problem) so that we get results back in the right order. This still results in only two queries, but we get the results back pre-sorted from the DB, so it should hopefully be a little faster (especially if we increase the max paging size).