Bypass ActiveRecord models for page queries
This is a little bit of a crazy optimization and we should not employ it lightly. However, the various page queries are *very* hot and just optimizing with Oj only gets us so far. ActiveRecord allocates a lot of not-quite-necessary objects, so this approach uses ActiveRecord to build a query, but then parses the results directly into a serializable JSON structure instead of allowing AR to create a whole pile of model objects. Since we have a simple JSON-like structure, we can then serialize it with Oj's strict mode, which is much faster than Rails mode (because it does not call `as_json` on any of the contained objects).

This approach *really* limits flexibility and is sensitive to lower-level changes in ActiveRecord, but it yields a 5x speed boost and a 4-5x memory savings, so seems worthwhile to use for the pages controller.