Add rake task for copying data from remote DBs
This uses the API to pull in pages and their versions from other databases (e.g. production or staging). They keep they same ID the had in the remote environment, so this can be really helpful for bringing in real data to test.

Right now, it only does one page at a time.

Usage:

    rake copy_page['https://api.monitoring.envirodatagov.org','YOUR USERNAME','YOUR PASSWORD','ID OF PAGE TO COPY']