Make diff endpoint's `data` the raw diff data
The idea here is that all the metadata and extra info the DB server provides should be in the standard `links` and `meta` sections of the response, while `data` should just be exactly what you'd get if you called the diff server directly. Makes a lot more sense in hindsight.

Fixes #181.