Switch to using aws-sdk-s3 from full aws-sdk
AWS has recently modularized the SDK gem so there is now one gem per service. Upgrading to V3 of the SDK is *huge*, so it's probably the right time to switch to a more trim and direct dependency on the S3 gem, which the only one we use. (Note that the major version for the full SDK gem is now 3.x, but the individual service gems are all new and at 1.x.)