POSTing a new, non-unique maintainer is an error
After a conversation with @danielballan, we decided POSTing is a request to create a new maintainer, and that should fail with an error ("409 Conflict" in this case) if a matching maintainer already exists.

When adding a maintainer to a page, that should be fine if it's already there, but should return the 409 error if it needed to create a new maintainer, but the new maintainer would have conflicted with an existing one (e.g. same name but different parent).