Consolidate to single docker file and make it run
This fixes the issues with the existing multi-stage files and combines the import worker and server definitions into a single Dockerfile.

Having both in one file allows us to reduce duplication without also adding some extra scripting around builds (otherwise we'd need three Dockerfiles and a script to make sure the base image is built before building the server or worker). This is less efficient than it could be, though -- Docker builds every stage in the file, but then only keeps the one you asked for. Ideally we'd solve that with a shell script, make, or gulp.