Increase redirect limit to 10 when archiving
Internet Archive stores all the redirects in the original request, which means they get replayed to us or anyone else who tries to retrieve an IA page. It turns out that sometimes that's a lot of redirects! Increase our limit to 10 for now (the default in HTTParty is 5).

This fixes some issues I hit while working on https://github.com/edgi-govdata-archiving/web-monitoring-processing/pull/174