Add retry/backoff to Archiver when downloading
When downloading raw content, the archiver should retry (up to two additional times) for gateway errors and timeouts. If the final result was still a gateway error, we process it as a success anyhow -- its possible that the actual snapshotted response was a gateway error and we don't want to fail to save that.

Fixes #261.