Add instructions for generating new seed data
I've been meaning to generate better seed data for a while (where most pages have multiple versions and where there are a fair number of pages), and finally did so for DCFemTech's hackathon this past weekend. I figured I'd add the script I wrote for it here so it's easy to run again if we need to (maybe to generate a different size import or just to create newer data).