Load test DB configuration more explicitly
OK, we've had a lot of back-and-forth and confusion over getting databases configured right when they require passwords or are running on different ports, etc. The idea here is that you can specify a special config by using the `DATABASE_URL` environment variable, which requires you to set all aspects of the connection. Additionally, this provides a workaround for the fact that Rails runs database rake commands twice with different configs if you are in the development environment.

If you are in a production or test environment, just set the `DATABASE_URL` environment variable.

If you are in a local environment, where you are using both development and test configurations, you can either:

- If you have the old-style setup with a Postgres DB requiring no authentication and superuser status, do nothing!
- Set `DATABASE_URL` to your development DB info in your `.env` (or `.env.development`) file and set `DATABASE_URL` to your test DB info in your `.env.test` file
- Set `DATABASE_URL` to your development DB info and `DATABASE_URL_TEST` to your test DB info

Also updates .gitignore to skip `.env.{environemnt name}` files (can't believe we didn't have this before!)

TODO: needs some README cleanup to make all the above more clear.
TODO: do we need to add more granular `DATABASE_*` items, like `DATABASE_USER`?