Stop archiving if the hash is known and archived
Since we store files by their hash, we know that a hash that is already in the store should be good, so we can stop early; there's [generally] no reason to download the content at all. If you *really* want to verify that the content at the URL to archive matches the already known hash, you can set the `force: true` argument.

This also updates the S3 storage layer to actually check whether the object is in the bucket (as opposed to just whether the bucket is correct). I think I had intended to do this before, but never got around to it! :(

Finally, the order of checks in `Archiver.already_archived?()` is reversed so that the cheapest check is first (there's no network hit for determining whether the URL is at an allowable, external archive host).

Fixes #377.