Add `?include_change_from_*` to version endpoints
Endpoints that return a version or a list of versions now support two additional query params:

- `?include_change_from_previous`
- `?include_change_from_earliest`

That include a field with the given name whose value the the change object (if any) representing the the change relative to the requested version. This is implemented *really* inefficiently (so it's very good that you have to ask for it), but I couldn't think of any way to do this better without adding a lot of denormalization. Regardless, I think it's still faster to have the API include this feature than to have the scraper make more API calls when building analyst task sheets.

This also depends on #406. I was hoping to avoid that, but big API changes here mean it intersects with that PR.