Factor priorities by percent change
This factors the source, text, and links components of our priority value by the percent change they represent. For example, a text change gets at least 0.1 priority, but it varies between 0.1 and 0.4 by the log of how much of the page's text was modified. Note we adjust the factor by the natural log -- there's probably as much room for messing with that formula as for the amount of priority allotted to any type of change.