Remove conflicting SSL config from initializers
The option this was setting is the default -- so it wasn't doing anything meaningful other than blowing away the config option that we were setting in the `production` environment config file.

This initializer just sets properties to their new defaults in Rails 5 (it's still hanging around from a very old upgrade) so you can test options that are about to change (or that just changed, depending on which side of the upgrade you're using it on). I'm pretty sure we could just delete this whole initializer file, but I think it'll be safer to save that for a separate commit.