Ensure version/page titles are nicely formatted
Version and Page titles should conform to some simple rules around formatting: namely that any surrounding spaces should be removed and that they are a single line. This also goes a step farther and normalizes repeated spaces (but not non-breaking spaces) into a single space, similar to HTML. This is mainly a backstop for import processes that might send us poorly formatted titles (perhaps by doing a poor job parsing HTML :(

I think I might also like to extend this concern to cover Tags and Maintainers in the future -- it seems reasonable that we'd want to apply the same rules to the `name` property on those models.

Fixes #430.