Add `active` boolean field to pages, fixes #432
This will be a manually-managed flag that helps us to decide whether to ask for and/or accept imports on certain pages. We have some URLs that we are tracking that either never really made sense to track or just don't really make sense to track anymore. This gives us a nice way to explicitly turn those off as we move away from Versionista as a system of record and merely as just a regular source of data.