Add `status` to pages and versions (#450, #451)
This adds the `status` field (a nullable int) to pages and versions, plus validations ensuring it is either `nil` or between 100 and 599 (inclusive). Imports automatically except any fields that are part of the version's schema and run validations, so this takes care of making sure the field is set properly on import for versions. Auto-updating the page's status will be handled separately because it's a much more complicated affair.