Change `Version#different` to ignore source_type
I originally thought it might be important to differentiate between the same content from multiple sources. That was probably wrong in retrospect, and especially so not that we are more actively using data from multiple sources and not attempting to differentiate much between them in pretty much any context.

Fixes #454.