Only auto-analyze versions that were different
If a version had the same content as the version that preceded it, then there's not a whole lot of point in auto-analyzing it -- there's nothing useful in the change between it and preceeding version or between it and the first version (because that change will be the same as between the preceeding version and the first version). Don't queue these versions for auto-analysis.