Fix seed.rb build (#563)
When we added the permissions array and removed the admin column, we forgot to update the seed script that checks for an admin user, and now it fails calling a method that no longer exists. This updates the seed script to check for a user with the `manage_users` permission.

Fixes #562.