Check `different` flag in `version.next/previous`
When looking for the `next` or `previous` version, we should almost always be looking for the next/previous version that was *different*. This was causing a lot of issues with calculating links for analysts, because new, non-different versions would be inserted after an analysis was done, and the analysis would be lost (because we'd try and retrieve it for the current vs. "previous" version, but "previous" was no longer the right thing).

This makes the default behavior one in which we check the different flag, but optionally lets you say `next(different: false)` to get the next version regardless of whether it's different.

Fixes #576.