Switch off `lightweight_query` by default
This is a first step to #348, removing `lightweight_query`. It adds an environment variable, `USE_LIGHTWEIGHT_QUERY`. If the variable is set, the old functionality will be used, but if not, it will default to normal ActiveRecord functionality. This way, we can have an emergency switch to flip it back on if things go wrong. In a later PR, we can remove the functionality (and this env var) entirely.

Note this has a kind of complicated way of accomplishing the `?include_versions` query. I couldn't figure out how to make it work in ActiveRecord as nicely as it does in `lightweight_query`, but I'm also not *super* worried about it since that whole thing is turned off and will probably be removed anyway.