Switch off `lightweight_query` by default (#582)
This is a first step to #348 (removing `lightweight_query`). It adds an environment variable, `USE_LIGHTWEIGHT_QUERY`. If the variable is set, the old `lightweight_query` functionality will be used, but if not, it will default to normal ActiveRecord functionality. This way, we can have an emergency switch to flip it back on if things go wrong. In a later PR, we can remove the functionality (and this env var) entirely.

Note this has a kind of complicated way of accomplishing the `?include_versions` query. I couldn't figure out how to make it work in ActiveRecord as nicely as it does in `lightweight_query`, but I'm also not *super* worried about it since that whole thing is turned off and will probably be removed anyway.

`lightweight_query` was a low-memory/CPU enhancement that can read an ActiveRecord relation object, run the query, and parse the results directly into an array of hashes. It avoided the overhead of instantiating a large set of ActiveRecord objects on every query. However, it also added a lot of complexity and brittleness to the system:
- It can only work with associations that would issue a single SQL query.
- It's sensitive to changes in ActiveRecord (it definitely has to reach into some less documented/not-explicitly-public bits)
- It makes `pages_controller` generally scary or confusing to work with.

Since we no longer have such extremely underpowered machines as we used to, we can function with the normal ActiveRecord overhead. Keeping the logic around for better performance isn't worth the tradeoffs in developer sanity.