Only include totals in list endpoints if `?include_total=true` (#596)
For large tables, counting the number of possible results (in order to determine the URL for the *last* chunk in a result set) turns out to be extremely expensive. (For example, in our production `versions` table, the main query takes < 20 ms, but counting the totals takes 18,000 ms.) Instead, only count the total results (`meta.total_results` in the response) and only include the URL for the last chunk (`links.last` in the response) if a user explicitly requests it using the `?include_total=true` query param.

In the future, we might move to an entirely different paging structure (that's not offset-based). See #579 for more.