Make sure `rake copy_page` copies ALL versions
We were failing to iterate through each chunk of versions, so if a page had > 100 versions, we were only copying the first 100! While I was at it, I added some summary statistics about versions that got skipped, and made sure the process stopped and failed if the page itself failed to copy.

Fixes #600.