Support credentials in env in copy_page rake task
When using `copy_page` to copy data from other API servers, you previously had to specify all your credentials as arguments, which means they'd show up the history, on the CLI when other people might be looking at your screen, etc. This change allows you to set those credentials in your environment so they don't show up on the command line itself and so that you don't have to keep re-typing them every time.

The environment variables are:
- WEB_MONITORING_DB_URL
- WEB_MONITORING_DB_EMAIL
- WEB_MONITORING_DB_PASSWORD

These are the same names used for the DB client class in the -processing project.

You can also set them in your `.env` file and they'll be picked up just fine.

NOTE: this also changes the order of CLI arguments! It is not backwards-compatible.