Remove dev/test gems from production Docker image
We were previously installing dev and test only gems in our production Docker image, which is less than ideal. This now installs only the production gems. It also creates a startup script for the API server, since we now need to precompile assets on startup instead of at container build time (the assets pipeline can read env vars, so you need a complete production configuration present, which might get baked into the pre-built assets, and we have none of that available until we deploy and run the images).