Format errors as JSON in dev mode (#615)
In dev mode, we used to show most errors with Rails's standard error page, which gives you some nice debugging tools. However, that's really not useful if you are testing a client against a local server. This change makes sure we format errors as JSON even in dev mode if the requester asks for JSON (either via the `Accept: application/json` header or the `.json` extension on a request).

This also adds stack traces to the error data in dev mode.

Fixes #614.