Calculate overall status code for a page (#619)
Pages have a `status` field, but until now we haven’t been filling it in. Because the occasional failure might happen when capturing a version, we don't want to just take the latest version's status. Instead, we only treat failures as real if they account for more than a certain percentage of the last N days.

`Page` objects now have an `update_status` method that will calculate and set the `status` field based on the above criteria. You can also query the set of pages that might need a status update using the `needing_status_update` scope:

    # Get all the NOAA pages that might need their status updated
    Page.needing_status_update.where('url LIKE ?', '%noaa.gov/%')

This also includes a Rake task to update the status of pages: `update_page_statuses`. By default, it updates pages that might need a status update based on the scope above. However, you can give it a single argument to work on a different selection of pages -- "all" for all pages, "unknown" for all pages with a NULL status, and "recent" for pages needing updates:

    $ bundle exec rake update_page_statuses[recent]

Finally, there is a new Docker image that runs the above task. We'll schedule it with Kubernetes so pages get regular status updates.