HOTFIX: Switch from readthis to built-in Redis cache (#636)
Rails 5.2 added a built-in Redis caching layer, but at the time, readthis was still maintained and there was no great reason to switch. However, Readthis has since declared they aren't updating it to support Rails 6+ and that the built-in one should be used instead, so this is required to deploy Rails 6.

AKA revenge of #378.