Add `/versions/{uuid}/raw` route (#475)
The new API route at `/versions/{uuid}/raw` returns the originally captured HTTP response body for a version. If the body is stored in a publicly available location, this serves a permanent redirect to that location.

This also changes the `get_file` method of `FileStorage::LocalFile` and `FileStorage::S3` to accept URLs and URIs in addition to the storage-relative paths it previously accepted.

95% of the work here is by @jsnshrmn; @Mr0grog tidied up around the edges.

Fixes #431.