Make seeding safe on badly configured machine
Running the seed script on a machine that wasn't properly configured with Resque, Redis, a diffing server, and an annotation user resulted in a whole mess of errors before. The basic idea here is twofold:

1. If you don't have things configured for auto-analysis, don't even try to schedule analysis jobs after importing.
2. If we can't connect to Redis and schedule auto-analysis, don't fail the import job. Just log a polite error message about it (in production, we should get a Sentry alert for this).

This should help ensure seeding succeeds when people are first setting things up and don't *also* have all the other stuff running yet.