Add `length` and `media_type` support for imports
The import job now sets `length` and `media_type` based on import record data. If not set, it will fall back to looking for `mime_type`, `encoding`, and `headers.Content-Length` in `source_metadta` for the relevant info (where we mostly send it now).

Finally, if the archiver is able to load the version's body (either remotely or if it already has it in storage), it will override whatever is set for `length` and replace it with the actual length. (This required adding `get_metadata()` to the file storage types. That's no major performance impact, though -- we already had to check whether the data existed in our storage.)