Perf: avoid unnecessary joins in pagination
It turns out we were issuing a really overcomplicated query when calculating pagination information for `Page` objects -- we'd join to maintainers and tags even if we didn't need to do any filtering on them, causing a two-step query for the relevant page IDs, rather than one. This two-step query has been KILLING our database in production.