Perf: avoid unnecessary joins in pagination (#687)
It turns out we were issuing a really overcomplicated query when calculating pagination information for `Page` objects -- we'd join to maintainers and tags even if we didn't need to do any filtering on them, causing a two-step query for the relevant page IDs, rather than one. This two-step query has been KILLING our database in production.

This only joins to those tables if we are actually filtering on them. It should drastically increase performance in our most common cases and not affect others.

This also flips from `eager_load` to `left_outer_joins` for this case. The `left_outer_joins` call used to be problematic, but no longer is. I think that’s because of how we split up the logic when we finally fetch the actual results, but may also be due to improvements in ActiveRecord between versions 5 and 6. Either way, hooray! 🎉