Don't rewrite lines in progress logs on non-TTYs
Without line endings, it's tough to stream logs (for example, we can't see the logs if we don't write line endings when running as a Kubernetes job). If running on a TTY, keep the friendly output that updates the line. On non-TTY interfaces, however, just write out the log as a line, and include a timestamp.