Allow versions to be orphaned from pages
We won't ever create an orphaned version situation without manual intervention (for now, at least). The main goal is to support a scenarios where we had two Pages:

    Page A=https://example.gov/a
    Page B=https://example.gov/b

1. Page B didn't originally exist, but was created when Page A was moved to that URL.
2. For a while, Page A redirected to Page B, versions attached to each of them have the same content.
3. Then Page A became a 404 page, so versions attached to A have different content than those attached to B.

A and B are the same conceptual page here, and now that we can support multiple URLs, we'd like to merge them. However, putting all the versions together would work fine right up until the point in time where A started responding with a 404 status code. Versions after that time will show up in our new, unified page's timeline as interleaved 200 and 404 responses, which is very confusing. To handle this, it should be possible to remove those 404 versions from the page. However, they have nowhere to go at this point, and no conceptual "page" that they really belong to.

(For now, at least, we won't be removing those versions from the page automatically. We might do so in the future.)