Delete merged pages, list in a table for redirects
Instead of leaving merged pages in place, delete them entirely and create records in a new `merged_pages` table that lists the old ID and the ID of the page they were merged into. Just in case we need it, the table has a JSON copy of most of the data from the old page.

I'm ever so slightly worried about whether this is the right move instead of just leaving the pages in places with a new column for merge targets. It's complex, but leaves us with fewer places to make mistakes if our indexes and queries aren't correctly excluding merged pages.