Compress responses with gzip or brotli (#859)
This adds middleware to compress responses with gzip or brotli encoding when clients accept it. Since JSON can be pretty repetitive, this saves a lot of bytes on the wire. We don't have a great place in our existing infrastructure to do this with a different way, so I think Rack middleware is probably the best approach for now.

Fixes #857.