Add API route for sampled list of versions
This adds a new API route at `/api/v0/pages/:page_id/versions/sampled` that gets a "sampled" list of versions -- that is, one version per sampling period. Right now, the sampling period is hard-coded to one day, but we might change that in the future. The main idea here is to give people a list of versions to look at, but reduce the amount of data in a useful way for pages that we capture extremely often (e.g. `https://epa.gov` gets many captures every day). For most use cases, we don't want to look more granularly than each day.

This is part of https://github.com/edgi-govdata-archiving/web-monitoring/issues/161