Paginate samples by date range (#992)
In #946, I added a /pages/:id/versions/sampled endpoint that returns no more than one page per date. The goal was to reduce the number of queries and the amount of data transferred to load some pages that have a lot of snapshots. It turns out this didn't work well because it still paginated by the number of versions it was querying, which means it still required the exact same number of requests (d'oh!). (It did reduce the amount of data sent, but not by much after gzipping responses. It was still many megabytes for some pages.)

This changes the approach to paginate by date range instead, which should make a big difference. The code is pretty rough, but as long as it works, I plan to deploy so we can try it out with the bigger production data set.