Only update the page title for non-error versions (#1061)
We want page titles to stay true to the conceptual content of the page, and when a page is removed, or when there are transient errors, we often see the page titles replaced by an error message. Instead, if a new version is an error, don't update the page record's title with the title from the error response. If there's no version with a valid title, use the filename component of the URL as a fallback.

This also adds some heuristics to determine whether pages with 200 (OK) status codes are actually errors.

Fixes #751, fixes #468.