Add new indexes for range-based pagination
Before we can actually do range-based pagination in any usefully workable way, we need to have indexes in place that support it (partially because the primary key for versions is a UUID, which is unordered -- this wouldn't be such a big deal with bigint). These indexes will take a while to build, so it's good to get these in place ahead of time.

A future change will remove the old, non-compound indexes that will be redundant after these new ones finish building.

Addresses part of #571, and required before deploying #1066.