Use range-based pagination for versions (#1066)
The slow performance of offset-based pagination (when you get a batch of results with a query like `SELECT xyz FROM table OFFSET = n LIMIT = m`) is a serious problem on the the `versions` table (really any large table, but in practice it’s just `versions` for this app). This solves the problem by using range-based pagination (skipping to the place you want in the result set using a `WHERE` clause that filters to a range of values from at the same field(s) you are sorting by).

For the versions table, that means doing queries based on a different minimum/maximum `(capture_time, uuid)` or `(created_at, uuid)` for each page/batch/chunk of results rather than using SQL's `OFFSET` clause to get to the chunk we want. This is more-or-less compatible with with any existing client, since the `next` link still works and the `chunk_size` parameter also works the same.

On the other hand, this introduces some new restrictions:

- You can only sort by one field, and it must be either `capture_time` or `created_at` instead of any field(s) you want. We need an index for the fields we want to sort by, and these two are the ones we chose.

- Clients shouldn’t usually be setting the `chunk` parameter on their own, and will run into problems if they do, since that field is no longer an integer (now it is either a version UUID or a combined `"<timestamp>,<uuid>"` string).

- The `?include_total=true` parameter is no longer supported for versions. There’s just no performant way to do it without additional caching of that field somewhere (`count()` in Postgres always requires a full table scan; the entire point of this change is to prevent us from doing large scans in order to paginate results).

There’s also plenty in here that’s not perfectly clean, and not totally generalized. If we wanted to utilize this kind of pagination for other models/controllers, it would need some work to make more abstract.