Handle bad credentials more gracefully
We previously required all requests to any API endpoint to be authenticated, so there was reason to differentiate between no credentials and invalid credentials. Now that we are enabling public view access, the difference matters. Devise doesn't provide anything that differentiates that way, so we have to drop down to check some Warden data.